refactor: `paths.d.ts` => `utils.d.ts`
